refactor: create ElmInlineIcon component and update imports in ElmJsonRenderer and ElmInlineText components
